修改：     Makefile 	修改：     boot/boot.S 	修改：     kernel/main.cpp 	修改：     kernel/mm/Page.h 	修改：     kernel/mm/mm.h

+++ kernel/mm/Page.h
@@ -62,7 +62,7 @@ void load_cr3(CR3 cr3) {
 
 #pragma pack(1)
 struct Info {
-	unsigned long long addr, size;
+	unsigned int addr, size;
 	unsigned int type;
 };
 union kit {
